util forLoop signature

+++ Data/List/RadixSort/Internal/LSD.hs
@@ -32,7 +32,7 @@ lsdRadixSort indexMap sortInfo @ SortInfo {..} digitsConstancy list@(x:_) =
         refVecFrom <- newSTRef vecIni
 
         M.when (siTopDigit > 0) $
-            forLoopM_ 1 (<= siTopDigit) (+1) $ \digit -> do
+            forLoopM_ (1::Int) (<= siTopDigit) (+1) $ \digit -> do
                 M.when ( not $ digitsConstancy V.! digit) $ digitPass digit refVecFrom
 
         readSTRef refVecFrom >>= V.unsafeFreeze >>= (return . collectVecToDList siTopDigitVal D.empty)
@@ -54,7 +54,7 @@ lsdRadixSort indexMap sortInfo @ SortInfo {..} digitsConstancy list@(x:_) =
                 vecTo <- VM.replicate (siTopDigitVal+1) S.empty
                 
                 -- readAndPartitionForLoop 0 (<= siTopDigitVal) (+1) vecFrom digit vecTo
-                forLoopM_ 0 (<= siTopDigitVal) (+1) $ \digitVal -> do
+                forLoopM_ (0::Int) (<= siTopDigitVal) (+1) $ \digitVal -> do
                     -- read vecFrom queue
                     s <- VM.unsafeRead vecFrom digitVal
                     -- partition to vecTo queues
